+++ R/options.R
@@ -2,8 +2,6 @@
   
   op <- options()
   op.searchConsoleR <- list(
-    searchConsoleR.client_id = "858905045851-3beqpmsufml9d7v5d1pr74m9lnbueak2.apps.googleusercontent.com",
-    searchConsoleR.client_secret = "bnmF6C-ScpSR68knbGrHBQrS",
     searchConsoleR.webapp.client_id = "858905045851-iuv6uhh34fqmkvh4rq31l7bpolskdo7h.apps.googleusercontent.com",
     searchConsoleR.webapp.client_secret = "rFTWVq6oMu5ZgYd9e3sYu2tm",
     searchConsoleR.scope = "https://www.googleapis.com/auth/webmasters",
@@ -26,3 +24,18 @@
   invisible()
   
 }
+
+
+.onAttach <- function(libname, pkgname){
+  
+  ## override any existing setting
+  options(googleAuthR.batch_endpoint = 'https://www.googleapis.com/batch/webmasters/v3',
+          googleAuthR.httr_oauth_cache = 'sc.oauth')
+  
+  suppressMessages(googleAuthR::gar_attach_auto_auth("https://www.googleapis.com/auth/webmasters", 
+                                    environment_var = "SC_AUTH_FILE"))
+  
+  invisible()
+  
+}
+